+++ src/Libraries/EDMS/Source/MODELS/robot.cc
@@ -59,7 +59,7 @@ extern void	( *idof_functions[MAX_OBJ] )( int ),
 
 //	Callbacks themselves...
 //	-----------------------
-extern void		( *EDMS_object_collision )( physics_handle caller, physics_handle victim, int badness, long DATA1, long data2, fix loc[3] ),
+extern void		( *EDMS_object_collision )( physics_handle caller, physics_handle victim, int32_t badness, int32_t DATA1, int32_t data2, fix loc[3] ),
 			( *EDMS_wall_contact )( physics_handle caller );
